# doc/api.yaml
openapi: 3.0.3 
info:
  title: WASAText api documentation
  description: Api documentation for the WASAText application
  version: '1'
servers:
  - url: http://localhost:3000
    description: Local server
...
paths:
...
  /chats/{chatId}:
    parameters:
        - name: chatId
          in: path
          required: true
          description: View one conversation
          schema:
            description: Chat id
            type: string
            pattern: '^.*?$' 
            minLength: 1
            maxLength: 999
            example: "abcdef012345"
    
    get:
      tags: ["Chats"]
      security:
        - bearerAuth: []
      summary: Lets the user view one specific conversation
      description: Given the conversation ID it will return all the messages in that conversation
      operationId: getConversation
      responses:
        '200':
          description: Fetched all messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/chat"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                description: error
                type: object
                properties:
                  errror:
                    description: error
                    type: string
                    pattern: '^.*?$'
                    minLength: 1
                    maxLength: 999
                    example: "Unauthorized"
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                description: error
                type: object
                properties:
                  errror:
                    description: error
                    type: string
                    pattern: '^.*?$'
                    minLength: 1
                    maxLength: 999
                    example: "Chat not found"
                       
    post:
      tags: [Messages]
      security:
        - bearerAuth: []
      summary: Send a message
      description: Sends the message to the selected chat
      operationId: sendMessage
      requestBody:
        description: Message
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/message"
      responses:
        '204':
          description: "Message sent succesfully"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                description: error
                type: object
                properties:
                  errror:
                    description: error
                    type: string
                    pattern: '^.*?$'
                    minLength: 1
                    maxLength: 999
                    example: "Unauthorized"
  /chats/{chatIdUrl}/messages/{messageIdUrl}:
    post:
      tags: [Messages]
      security:
        - bearerAuth: []
      summary: Forwards a message
      description: Forwards the selected message to a new chat
      operationId: forwardMessage
      parameters:
        - name: chatIdUrl
          description: Chat id
          in: path
          required: true
          schema:
            description: Chat id
            type: string
            example: "chat3762"
            pattern: '^.*?$'
            minLength: 1
            maxLength: 999
        - name: messageIdUrl
          description: Message id
          in: path
          required: true
          schema:
            description: Message id
            type: string
            example: "message5763"
            pattern: '^.*?$'
            minLength: 1
            maxLength: 999
      requestBody:
        description: Select chat to forward to
        content:
          application/json:
            schema:
              description: Chat to forward to
              type: object
              properties:
                messageId:
                  description: Message id
                  type: string
                  example: "message4357"
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 999
                chatId: 
                  description: Chat id
                  type: string
                  example: "chat349875"
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 999
      responses:
        '204':
          description: "Message forwarded"
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                description: error
                type: object
                properties:
                  errror:
                    description: error
                    type: string
                    example: "Unauthorized"
                    pattern: '^.*?$'
                    minLength: 1
                    maxLength: 999
        '404':
          description: "Message or chat not found"
          content:
            application/json:
              schema:
                description: error
                type: object
                properties:
                  errror:
                    description: error
                    type: string
                    example: "Message or chat not found"
                    pattern: '^.*?$'
                    minLength: 1
                    maxLength: 999
...
  /chats/{chatId}/messages/{messageId}/delete:
    delete:
      tags: [Messages]
      security:
        - bearerAuth: []
      summary: Deletes a message
      description: Deletes the selected message as long as it was sent by the user
      operationId: deleteMessage
      parameters:
        - name: chatId
          description: Chat id
          in: path
          required: true
          schema:
            description: Chat id
            type: string
            example: "chat3762"
            pattern: '^.*?$'
            minLength: 1
            maxLength: 999
        - name: messageId
          description: Message id
          in: path
          required: true
          schema:
            description: Message id
            type: string
            example: "message5763"
            pattern: '^.*?$'
            minLength: 1
            maxLength: 999
      responses:
        '204':
          description: Message deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                description: error
                type: object
                properties:
                  errror:
                    description: error
                    type: string
                    example: "Unauthorized"
                    pattern: '^.*?$'
                    minLength: 1
                    maxLength: 999
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                description: error
                type: object
                properties:
                  errror:
                    description: error
                    type: string
                    example: "Message not found"
                    pattern: '^.*?$'
                    minLength: 1
                    maxLength: 999
...
  /createchat:
    post:
      tags: ["Chats"]
      security:
        - bearerAuth: []
      summary: Create a new chat
      description: Create a new chat
      operationId: createChat
      requestBody:
        description: Chat details
        content:
          application/json:
            schema:
              description: Chat details
              type: object
              properties:
                chatName:
                  description: Chat name
                  type: string
                  example: "University chat"
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 100
                chatType:
                  description: Chat type
                  type: string
                  enum:
                    - "group"
                    - "private"
                  example: "group"
                chatParticipants:
                  description: Chat participants
                  type: array
                  minItems: 1
                  maxItems: 999
                  items:
                      # description: user
                      # type: integer
                      # example: "1"
                    $ref: "#/components/schemas/user"
        required: true
      responses:
        '201':
          description: Chat created
          content:
           application/json:
            schema:
              description: Chat id
              type: object
              properties:
                chatId:
                  description: Chat id
                  type: string
                  example: "chat37543"
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 999
components:
  schemas:
    message:
      description: Message
      type: object
      properties:
        text:
          description: Message text
          type: string
          example: "Ciao"
          pattern: '^.*?$'
          minLength: 1
          maxLength: 9999
        photo:
          description: Message photo
          type: string
          format: binary
          minLength: 0
          example: ""
        sender: 
          description: Sender id
          type: integer
          minimum: 1
          maximum: 999
          example: 1
        receiver:
          description: Receiver id
          type: integer
          minimum: 1
          maximum: 999
          example: 3
        dateTime:
          description: Date and time of the message
          type: string
          format: date-time
          example: "2024-12-02T16:38:00"
        reactions:
          description: Reactions to the message
          type: array
          minItems: 0
          maxItems: 999
          items:
            type: object
            properties:
              utente:
                description: User name
                type: string
                example: "William"
                pattern: '^.*?$'
                minLength: 3
                maxLength: 16
              reazione:
                description: Reaction
                type: string
                example: "emoji"
                minLength: 0
                maxLength: 1
    user:
      type: object
      properties:
        userId:
          type: integer
          minimum: 1
          maximum: 999
          example: 1
        userName:
          type: string
          example: "luiif"
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
        userPhoto:
          type: string
          format: binary
          minLength: 0
          example: ""
    chat:
      description: Chat
      type: object
      required:
        - chatType
      properties:
        chatId:
          description: Chat id
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 999
          example: "chat37543"
        chatName:
          description: Chat name
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 100
          example: "The salami 2 the return"
        chatPhoto:
          description: Chat photo
          type: string
          format: binary
          minLength: 0
          example: ""
        chatType:
          description: Chat type
          type: string
          enum:
            - "group"
            - "private"
          example: "group"
        chatParticipants:
          description: Chat participants
          type: array
          minItems: 1
          maxItems: 999
          items:
            $ref: "#/components/schemas/user"
        chatMessages:
          description: Chat messages
          type: array
          minItems: 0
          maxItems: 999
          items:
            $ref: "#/components/schemas/message"
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: apiKey
      in: header
      name: Authorization